# apps/backend/docs/openapi/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_arc-platform_backend_modules_compliance_service.ConsentBasis:
    enum:
    - explicit
    - legitimate_interest
    - contractual
    - legal_obligation
    type: string
    x-enum-varnames:
    - ConsentBasisExplicit
    - ConsentBasisLegitimateInterest
    - ConsentBasisContractual
    - ConsentBasisLegalObligation
  github_com_arc-platform_backend_modules_compliance_service.ConsentRequest:
    properties:
      asset_id:
        type: string
      consent_basis:
        $ref: '#/definitions/github_com_arc-platform_backend_modules_compliance_service.ConsentBasis'
      consent_expires_at:
        type: string
      consent_obtained_at:
        type: string
      metadata:
        additionalProperties: true
        type: object
      obtained_by:
        type: string
      pii_type:
        type: string
      purpose:
        type: string
    required:
    - asset_id
    - consent_basis
    - consent_obtained_at
    - obtained_by
    - pii_type
    - purpose
    type: object
  github_com_arc-platform_backend_modules_compliance_service.ConsentWithdrawalRequest:
    properties:
      withdrawal_reason:
        type: string
      withdrawal_requested_by:
        type: string
    required:
    - withdrawal_requested_by
    type: object
...
host: localhost:8080
info:
  contact:
    email: security@arc-hawk.io
    name: ARC-Hawk Security Team
  description: Enterprise PII Discovery, Classification, and Lineage Tracking Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://arc-hawk.io/terms
  title: ARC-HAWK-DD API
  version: 3.0.0
paths:
...
  /api/v1/consent/records:
    get:
      consumes:
      - application/json
      description: Returns consent records with optional filters
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by asset ID
        in: query
        name: asset_id
        type: string
      - description: Filter by PII type
        in: query
        name: pii_type
        type: string
      - description: Filter by consent status
        in: query
        name: status
        type: string
      - description: Maximum records to return
        in: query
        name: limit
        type: integer
      - description: Number of records to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List consent records
      tags:
      - compliance
    post:
      consumes:
      - application/json
      description: Records a new consent entry for a data subject
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Consent request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_arc-platform_backend_modules_compliance_service.ConsentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Record a new consent
      tags:
      - compliance
  /api/v1/consent/status/{assetId}/{piiType}:
    get:
      consumes:
      - application/json
      description: Returns the consent status for a specific asset and PII type combination
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Asset ID
        in: path
        name: assetId
        required: true
        type: string
      - description: PII type
        in: path
        name: piiType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get consent status for an asset and PII type
      tags:
      - compliance
  /api/v1/consent/violations:
    get:
      consumes:
      - application/json
      description: Returns assets that have consent violations
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get consent violations
      tags:
      - compliance
  /api/v1/consent/withdraw/{id}:
    post:
      consumes:
      - application/json
      description: Withdraws a previously recorded consent by ID
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Consent ID
        in: path
        name: id
        required: true
        type: string
      - description: Withdrawal request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_arc-platform_backend_modules_compliance_service.ConsentWithdrawalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Withdraw an existing consent
      tags:
      - compliance
...
securityDefinitions:
  APIKeyAuth:
    description: Tenant API key for service-to-service authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"